tokenpro/requirements.txt
asgiref==3.3.4
Django==3.2
djangorestframework==3.12.4
dnspython==2.1.0
PyJWT==2.0.1
pymongo==3.11.3
pytz==2021.1
sqlparse==0.4.1
